+++ Sources/Spectre/Global.swift
@@ -6,6 +6,8 @@ import CRT
 import Darwin
 #endif
 
+import Foundation
+
 
 let globalContext: GlobalContext = {
 #if os(macOS)
@@ -25,6 +27,39 @@ public func it(_ name: String, _ closure: @escaping () throws -> Void) {
   globalContext.it(name, closure: closure)
 }
 
+#if swift(>=4.2)
+enum ReporterType: String, CaseIterable {
+  case tap
+  case dot
+}
+#endif
+
+fileprivate func defaultReporter() -> Reporter {
+  if let reporterEnv = ProcessInfo.processInfo.environment["SPECTRE_REPORTER"] {
+#if swift(>=4.2)
+    guard let reporterType = ReporterType(rawValue: reporterEnv) else {
+      let supported = ReporterType.allCases.map { $0.rawValue }
+      let error = "Unknown reporter: \(reporterEnv). Supported: \(supported.joined(separator: ", "))\n"
+      FileHandle.standardError.write(error.data(using: .utf8)!)
+      exit(4)
+    }
+
+    switch reporterType {
+    case .tap:
+      return TapReporter()
+    case .dot:
+      return DotReporter()
+    }
+#else
+    let error = "SPECTRE_REPORTER is unsupported on Swift < 4.1\n"
+    FileHandle.standardError.write(error.data(using: .utf8)!)
+    exit(4)
+#endif
+  }
+
+  return StandardReporter()
+}
+
 public func run() -> Never  {
   let reporter: Reporter
 
@@ -33,7 +68,7 @@ public func run() -> Never  {
   } else if CommandLine.arguments.contains("-t") {
     reporter = DotReporter()
   } else {
-    reporter = StandardReporter()
+    reporter = defaultReporter()
   }
 
   run(reporter: reporter)